fix: movendo requirements.txt e main.py para raiz
